requirements.txt
Flask
sentence-transformers
pinecone-client
Werkzeug==2.2.2
numpy